LayoutTests/platform/blackberry/fast/filesystem/simple-persistent-expected.txt

CONSOLE MESSAGE: line 24: ReferenceError: Can't find variable: webkitRequestFileSystem
FAIL: Timed out waiting for notifyDone to be called
requestFileSystem PERSISTENT test.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".
